+++ tests/zar/test_cli_unzar.py
@@ -1,4 +1,3 @@
-import logging
 import os
 import shutil
 from typing import List
@@ -6,7 +5,8 @@ from typing import List
 from tests.zar import check_dir_and_remove, check_zip_and_remove, test_directory, test_files
 from zmxtools import cli
 
-log = logging.getLogger(__name__)
+from tests import log
+log = log.getChild(__name__)
 
 
 def test_unzar_basic():